refactor: Organize pkg into generic digit_span_task
