+++ lib/digit_span_tasks.dart
@@ -1,5 +1,5 @@
 library digit_span_tasks;
 
-export 'src/digit_span_tasks_activity.dart';
-export 'src/components/data/digit_span_tasks_data.dart';
-export 'src/components/config/user_config.dart';
+export 'src/digit_span_task/digit_span_tasks_activity.dart';
+export 'src/digit_span_task/components/data/digit_span_tasks_data.dart';
+export 'src/digit_span_task/components/config/user_config.dart';
