+++ lib/src/digit_span_task/components/data/trial_data.dart
