+++ lib/src/digit_span_task/components/end/end_view.dart
@@ -1,0 +1,32 @@
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+import 'package:digit_span_tasks/src/digit_span_task/components/ui_components/default_text.dart';
+import 'package:digit_span_tasks/src/digit_span_task/components/end/end_controller.dart';
+import 'package:digit_span_tasks/src/digit_span_task/components/ui_components/screen.dart';
+
+/// Screen for instructing the participant the task has ended.
+/// Doesn't allow going back or restarting so the experimenter has control
+/// over the session.
+class EndView extends StatelessWidget {
+  EndView({Key? key}) : super(key: key);
+  final EndController _endController = Get.put(EndController());
+
+  @override
+  Widget build(BuildContext context) {
+    _endController.toNextScreen();
+    return WillPopScope(
+      onWillPop: () async {
+        return false;
+      },
+      child: Screen(
+        children: Column(
+          children: const <Widget>[
+            DefaultText(
+              '¡Terminamos esta fase!',
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
