+++ lib/src/digit_span_task/components/errors/errors.dart
