+++ lib/src/digit_span_task/components/ui_components/default_text.dart
